refactor multiplexer mode (#497)
