+++ .github/workflows/testing-go.yml
@@ -23,7 +23,7 @@ jobs:
       matrix:
         os-version: ['ubuntu-22.04', 'macos-latest' ]
         go-version: [ '1.20', '1.21' ]
-        package: ['dnsutils', 'collectors', 'loggers', 'transformers', 'netlib', 'processors']
+        package: ['dnsutils', 'collectors', 'loggers', 'transformers', 'netlib', 'processors', '.']
         exclude:
           - os-version: macos-latest
             go-version: '1.20'
