+++ Makefile
@@ -57,6 +57,7 @@ tests:
 	@go test -timeout 30s ./collectors/ -race -cover -v
 	@go test -timeout 90s ./loggers/ -race -cover -v
 	@go test -timeout 30s ./processors/ -race -cover -v
+	@go test -timeout 30s . -race -cover -v
 
 clean:
 	@go clean
